rename: types폴더 공통화
