+++ src/components/Editor/RichEditor.tsx
@@ -9,7 +9,7 @@ import ReactQuill, { Quill } from 'react-quill';
 
 import { IMAGE_EXTENSION, IMAGE_TYPE } from '@/constants';
 import useImageUpload from '@/hooks/common/useImageUpload';
-import { SubTravelogueType } from '@/types/post';
+import { SubTravelogueType } from '@/types/travelogue';
 
 import { editorModules } from './index';
