+++ src/components/Travelogue/Feed/Feed.tsx
@@ -2,7 +2,7 @@ import { Stack } from '@mui/material';
 import { useRouter } from 'next/router';
 
 import useAuth from '@/hooks/auth/useAuth';
-import { TravelogueFeedType } from '@/types/travelogue';
+import { TravelogueFeedType } from '@/types/main';
 import { setItem } from '@/utils/storage';
 
 import { FeedContent, FeedHeader, FeedImage } from '.';
