+++ src/components/common/FileInput.tsx
@@ -5,7 +5,7 @@ import { useState } from 'react';
 import { ControllerRenderProps } from 'react-hook-form';
 
 import useImageUpload from '@/hooks/common/useImageUpload';
-import { TravelogueFormType } from '@/types/post';
+import { TravelogueFormType } from '@/types/travelogue';
 
 import { SubTitle } from './';
