+++ src/hooks/query/member.ts
@@ -6,7 +6,7 @@ import { getUserInformation, patchUserInformation } from '@/api/member';
 import { NICKNAME_PATTERN } from '@/constants/pattern';
 import useImageUpload from '@/hooks/common/useImageUpload';
 import { userInformationState } from '@/recoil';
-import { UserInformationPatchType } from '@/types/profile';
+import { UserInformationPatchType } from '@/types/member';
 
 export const useUserInformation = () => {
   const [userInformation, setUserInformation] = useRecoilState(userInformationState);
