+++ src/pages/detail/index.tsx
@@ -5,7 +5,7 @@ import React, { useEffect, useState } from 'react';
 import Spinner from '@/components/common/Spinner';
 import { TravelogueDetail } from '@/components/TravelogueDetail';
 import { usePatchTravelogueDetailById } from '@/hooks/query/detail';
-import { TravelogueDetailType } from '@/types/travelogue';
+import { TravelogueDetailType } from '@/types/detail';
 
 const DetailPage = () => {
   const router = useRouter();
