+++ src/types/search.ts
